requirements.txt
Flask==2.0.1
python-dotenv==0.19.0
requests==2.26.0
responses==0.14.0
authlib>=0.14.1
python-jose
flask-cors
six
flask-login